setup auth and users module
