+++ apps/backend/src/authentication/authentication.service.ts
@@ -5,7 +5,7 @@ import {
 } from '@nestjs/common';
 
 import { randomBytes, scrypt as _scrypt } from 'crypto';
-import { UsersService } from 'src/users/users.service';
+import { UsersService } from '../users/users.service';
 import { promisify } from 'util';
 
 const scrypt = promisify(_scrypt);
