+++ apps/backend/src/users/dtos/create-user.dto.ts
@@ -5,4 +5,6 @@ export class CreateUserDto {
   email: string;
   @IsString()
   password: string;
+  @IsString()
+  confirmPassword: string;
 }
