+++ apps/backend/src/users/users.controller.ts
@@ -1,64 +1,22 @@
 import {
   Body,
   Controller,
-  Post,
   Get,
   Param,
   Query,
   Delete,
   Patch,
   NotFoundException,
-  Session,
-  UseGuards,
 } from '@nestjs/common';
-import { CreateUserDto } from './dtos/create-user.dto';
 import { UsersService } from './users.service';
 import { UpdateUserDto } from './dtos/update-user.dto';
 import { Serialize } from '../interceptors/serialize.interceptor';
 import { UserDto } from './dtos/user.dto';
-import { CurrentUser } from './decorators/current-user.decorator';
-import { User } from './user.entity';
-import { AuthenticationService } from 'src/authentication/authentication.service';
-import { AuthGuard } from '../guards/AuthGuard';
 
-@Controller('auth')
+@Controller('users')
 @Serialize(UserDto)
-// @UseInterceptors(CurrentUserInterceptor) // controllers scoped interceptor
 export class UsersController {
-  constructor(
-    private usersService: UsersService,
-    private authService: AuthenticationService,
-  ) {}
-
-  // @Get('/whoami')
-  // async whoAmI(@Session() session: any) {
-  //   return await this.usersService.findOne(session.userId);
-  // }
-
-  @Get('/whoami')
-  @UseGuards(AuthGuard)
-  async whoAmI(@CurrentUser() user: User) {
-    return user;
-  }
-
-  @Post('/signup')
-  async createUser(@Body() body: CreateUserDto, @Session() session: any) {
-    const user = await this.authService.userSignUp(body.email, body.password);
-    session.userId = user.id;
-    return user;
-  }
-
-  @Post('/signin')
-  async signInUser(@Body() body: CreateUserDto, @Session() session: any) {
-    const user = await this.authService.userSignIn(body.email, body.password);
-    session.userId = user.id;
-    return user;
-  }
-
-  @Post('/signout')
-  async signOutUser(@Session() session: any) {
-    session.userId = null;
-  }
+  constructor(private usersService: UsersService) {}
 
   @Get('/:id')
   async findUser(@Param('id') id: string) {
