+++ apps/backend/src/users/users.module.ts
@@ -1,12 +1,9 @@
 import { MiddlewareConsumer, Module } from '@nestjs/common';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthenticationModule } from 'src/authentication/authentication.module';
 import { CurrentUserMiddleware } from './middlewares/current-user.middleware';
 import { User } from './user.entity';
-import { AuthenticationService } from 'src/authentication/authentication.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
